funções de somar e subtrair a quantidade com base nas entradas e saídas de produtos

+++ ibell-embalagens-api/src/main/java/br/com/uniamerica/ibellembalagens/Entity/Product.java
@@ -20,7 +20,7 @@ public class Product extends AbstractEntity {
 
     @Getter @Setter
     @Column(name = "quantity", length = 10, nullable = false)
-    private BigDecimal quantity;
+    private Float quantity;
 
     @Getter @Setter
     @Column(name = "unit_value", length = 10, nullable = true)
